DISTX-112. Update user sync API
The existing user sync API has been moved to "create". This will
be left in place until the user sync API work is completed.

The user sync API has been changed to accept an optional set of
environments and users in anticipation of retrieving users from
the UMS instead of passing them in through the API. The environments
and users in the request will be used to narrow the scope of the
sync. e.g., if we want to sync only one environment or one user.

Minor improvements were made to API documentation and testing.